+++ types/beatmap/wrapper/mod.ts
@@ -4,19 +4,22 @@
  */
 
 export * from './arc.ts';
-export * from './baseObject.ts';
+export * from './baseItem.ts';
+export * from './baseNote.ts';
 export * from './baseSlider.ts';
 export * from './bombNote.ts';
 export * from './bpmEvent.ts';
 export * from './chain.ts';
 export * from './colorBoostEvent.ts';
 export * from './colorNote.ts';
+export * from './container.ts';
 export * from './difficulty.ts';
 export * from './event.ts';
 export * from './eventBox.ts';
 export * from './eventBoxGroup.ts';
 export * from './eventTypesForKeywords.ts';
 export * from './eventTypesWithKeywords.ts';
+export * from './gridObject.ts';
 export * from './indexFilter.ts';
 export * from './lightColorBase.ts';
 export * from './lightColorEventBox.ts';
@@ -24,6 +27,9 @@ export * from './lightColorEventBoxGroup.ts';
 export * from './lightRotationBase.ts';
 export * from './lightRotationEventBox.ts';
 export * from './lightRotationEventBoxGroup.ts';
+export * from './lightTranslationBase.ts';
+export * from './lightTranslationEventBox.ts';
+export * from './lightTranslationEventBoxGroup.ts';
 export * from './obstacle.ts';
 export * from './rotationEvent.ts';
 export * from './waypoint.ts';